Changed NeighborhoodRule.magnitude to isize
Changed magnitude for an isize & used sign to indicate direction, obviating the need for
EdgeDirection enum.

While implementing neighborhood rules in the runtime, it became clear that this was a simpler
solution because coordiantes are isizes.

Coordiantes must be an isize to allow for infinite boundaries; they must be able to grow
in any direction, positive or negative.