Implemented prototype 2D renderer
Implemented a prototype for a renderer for 2D simulations.

This version generates frames & writes them to disk, where the can be
stiched together with `ffpmeg` or `mencoder`. It would be prefferable to
use the `ffpmeg` crate to do this inside of Loaf without involving
another command.

Procedure for generating videos:

http://electron.mit.edu/~gsteele/ffmpeg/