Total rewrite of project
After inadvertently putting this repo into a bad state, I decided it would
be better to rewrite the project rather than recover the files.

During the first attempt, I was unsure of how to implement a programming
language. I started writing the parser first, because this was the component
I was least sure of the approach for. This resulted in an architecture that
was oriented to serve the parser, and made implementing the runtime awkward.

When I rewrote the project, I started from the runtime out (in fact, I
haven't rewritten the parser or 2D renderer yet). This allowed me to focus
on Loaf's core competancy - running Loaf code - and rest of the architecture
decisions will follow from this core.

I have implemented neighorhood and state transition rules, a naive fixed-size
grid environment, and a synchronous runtime. I have written an implementation
of Conway's Game of Life using these primitives, and verified it against a
few trivial patterns.

Regrettably, I do not have a git history for the rewrite up to this point;
I wrote it in another project, and simply chipped away at it until it was
done. Going forward I will record commits for incremental changes.